+++ knodle/trainer/crossweigh_weighing/dscrossweigh.py
@@ -11,7 +11,7 @@ from torch.utils.data import TensorDataset
 
 from knodle.trainer.baseline.majority import MajorityVoteTrainer
 from knodle.trainer.crossweigh_weighing.config import DSCrossWeighDenoisingConfig
-from knodle.trainer.crossweigh_weighing.sample_weights_calculator import DSCrossWeighWeightsCalculator
+from knodle.trainer.crossweigh_weighing.dscrossweigh_weights_calculator import DSCrossWeighWeightsCalculator
 
 from knodle.transformation.filter import filter_empty_probabilities
 from knodle.transformation.majority import z_t_matrices_to_majority_vote_probs
