+++ tests/trainer/crossweigh/test_cw_labels_calculation.py
@@ -4,7 +4,7 @@ import torch
 from torch.utils.data import TensorDataset
 
 from knodle.model.logistic_regression_model import LogisticRegressionModel
-from knodle.trainer.crossweigh_weighing.sample_weights_calculator import DSCrossWeighWeightsCalculator
+from knodle.trainer.crossweigh_weighing.dscrossweigh_weights_calculator import DSCrossWeighWeightsCalculator
 
 
 @pytest.fixture(scope='session')
